% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/db.R
\name{orderly_db_query}
\alias{orderly_db_query}
\title{Extract data from a database}
\usage{
orderly_db_query(query, as, database = NULL)
}
\arguments{
\item{query}{Query to evaluate}

\item{as}{Name of the object to create}

\item{database}{The name of the database. This can be omitted (or
\code{NULL}) where you only have a single database, but must be
specified if you have more than one database configured.}
}
\value{
Undefined
}
\description{
Extract data from a database
}
